playbooks: split per "role" (actually host)
Docs "Best practices" say to split per "role".
In our case each host is different.

This allows to list the tasks to set up a host using --list-tasks.
--check mode doesn't work, if you have tasks that depend on others
This includes if you set up a minimal chroot and then try to
install packages.  The Ansible apt module tries to automatically
install python-apt, but of course it's not allowed to do so in
check mode.