Fixes inc. etckeeper when `dnsdomainname` is empty (nspawn)
Inside systemd-nspawn OS container, `dnsdomainname` returns empty.
git becomes unhappy with the email address it guesses for us.
When git is run from ansible (but not the command line),
it treats this as a fatal error.

Set user.email explicitly.  It will be useless a lot of the time,
but we don't really care.  This is only set on the etckeeper repo,
so it doesn't affect any other uses of git by the root user.