roles/debian-on-nspawn-dhcp using ifupdown
Sadly it takes a few seconds...
this is actually due to the DHCP server:
it only happens on fedora where I use libvirt network,
it does not happen on a debian jessie system which
just uses systemd-networkd on the host side.

Google suggests this might be due to the DHCP server using ping.
(It's certainly familiar that dnsmasq setups can seem to
take several seconds).