Clarify roles which are definitely Debian-specific
The smokeping role includes Debian-specific code.  It works around the
Jessie package not dropping in to apache.  Presumably this will have to
be removed in future, plus a caveat.

The nspawn support is intended to work on Debian or Ubuntu, which I
indicate with the prefix "deb-" instead.

In general, the roles that aren't specifically written to be portable will
almost certainly only work on Debian-based distributions.  I might rename
some more later.  Maybe this didn't help in itself.   But, I definitely
want to prefix my role for setting up apache+ssl, so I was looking for the
pattern here.