brick upgrade to stretch
manually deleted sources.list.d/*

stretch-srv suffered unintended relabelling and was not recovered.
(btw `man systemd-nspawn` says there's a way to do it).
Instead apt-cacher-ng was moved to brick, which works for now
(i.e. until one wants to test a newer version of Debian and apt-cacher-ng
is incompatible with it).