roles.srv/ for server-specific stuff
roles/ is very cluttered.  It's difficult to pick things out when it
contains all of common roles, server roles, and a few GUI-specific roles.

So at minimum I want to splt out common or server roles.  Spliting out
common on it's own kinda begs the question of what the different non-common
cases are.  So it's easiest to just split out the server roles.

(In principle this suggests the few GUI roles should be split out as well.
But there's only really 3, so in practice it seems easier not to bother).