roles: daily backups: switch back from firefox profile to just bookmarks
Tested firefox profile added 50M to bup daily, 180M to borg daily.

For bup, this was quite concerning, it is 1/75th of the tested home
directory (dailies without the profile were 10x lower or less).

borg is better in the sense that I could choose to prune down e.g. to less
than 25 snapshots if I wanted.  But.  That is not a great number of
snapshots.  And I am not really happy to have Firefox taking up half
the backup... and being nastily fragmented at the same time.  The backups
are still very useful without Firefox.

That said, I am attached to my bookmarks.  And Firefox Sync doesn't protect
against malice or equivalent.  So I want to keep backing those up.

The original bookmarkbackups code from daily-bup is replaced, with
a more generic approach.  I.e. it demands less from the backup application.
This places the bookmarks in a separate named backup archive.

If I ever manage to generalize the backup roles further, this may merit
replacement again.  `.mozilla/firefox/*/bookmarkbackups` requires
supporting wildcards.  It would be more general to support pre-scripts,
which could copy the bookmarkbackups to a staging area.  Similar to how
a pre-script could be used to snapshot a filesystem or database
(potentially taking advantage of their crash safety).