roles: borgbackup: enable compression
LZ4 is wicked fast, it did not appear to slow down create (on a desktop
celeron).  It was also halving the daily size.  (It works even better on
the excluded index DBs.

The block sizes still make it worse than bup even with zlib.  bup dailys
are still coming out half the size after all this, and they _include_
the index DBs.  But bup wasn't really designed for pruning; the performance
there is abysmal.