jenkins-desktop: /home/jenkins-photos -> /home/shared/Photos
It *sounds* so silly that we didn't also have a /home/shared for shared
writable documents.  But the problem is to make sure everything still gets
backed up as we want it to.

Also /home/music.cd -> /home/shared/Music, making things look a bit more
consistent.

Music and Photos are backed up separately.  Shared documents are expected
to be a different size class, so may be distributed different when making
extra replica backups.

* roles/jenkins-photos
* roles/daily-bup
* roles/daily-borgbackup
* playbooks/jenkins-desktop.yml

The last is because borgbackup is now finally marked as the primary local
backup.