roles: packages-cli-basics
design:
- ensure you have certain programs when you need them (outages etc).
- "not very large" etc
- x-distro

The list is chosen to match my taste, though I think a majority
are unobjectionable.

So this ends up being constrained by the x-distro thing.
It would probably prefer to have separate list for Debian and Fedora.

This might be particularly useful for a range of Debian
install methods.  I think the "Debian standard" live image
omitted almost all of these.